chore(fe-02): move price-related float fields to int cent-based
Move price-related float fields to int cent-based
Update gems
Fix problem with installments_count during hotwire update
Fix bug regarding pressing ENTER on input during hotwire update

+++ db/migrate/20231206000007_create_money_transactions.rb
@@ -9,8 +9,8 @@ class CreateMoneyTransactions < ActiveRecord::Migration[7.1]
       t.date :date, null: false
       t.integer :month, null: false
       t.integer :year, null: false
-      t.decimal :starting_price, null: false
-      t.decimal :price, null: false
+      t.integer :starting_price, null: false
+      t.integer :price, null: false
       t.boolean :paid, default: false
       t.string :money_transaction_type, null: true
