+++ db/migrate/20231206000008_create_card_transactions.rb
@@ -9,8 +9,8 @@ class CreateCardTransactions < ActiveRecord::Migration[7.1]
       t.date :date, null: false
       t.integer :month, null: false
       t.integer :year, null: false
-      t.decimal :starting_price, null: false
-      t.decimal :price, null: false
+      t.integer :starting_price, null: false
+      t.integer :price, null: false
       t.integer :installments_count, default: 0, null: false
 
       t.references :user, null: false, foreign_key: true
